+++ pkg/hypershift/hypershift.go
@@ -7,39 +7,27 @@ import (
 
 	"github.com/stolostron/cm-cli/pkg/genericclioptions"
 	"github.com/stolostron/cm-cli/pkg/helpers"
+	hypershiftdeploymentv1alpha1 "github.com/stolostron/hypershift-deployment-controller/api/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/kubectl/pkg/cmd/get"
-	clusterv1 "open-cluster-management.io/api/cluster/v1"
 )
 
-const (
-	ConsoleURLClusterClaim string = "consoleurl.cluster.open-cluster-management.io"
-)
-
-func OpenHosted(cmFlags *genericclioptions.CMFlags, clusterName string, timeout int, printFlags *get.PrintFlags) error {
+func GetHypershiftDeployment(clusterName string, cmFlags *genericclioptions.CMFlags) (*hypershiftdeploymentv1alpha1.HypershiftDeployment, error) {
 	dynamicClient, err := cmFlags.KubectlFactory.DynamicClient()
 	if err != nil {
-		return err
+		return nil, err
 	}
-	mcu, err := dynamicClient.Resource(helpers.GvrMC).Get(context.TODO(), clusterName, metav1.GetOptions{})
+
+	hdus, err := dynamicClient.Resource(helpers.GvrHD).List(context.TODO(), metav1.ListOptions{
+		LabelSelector: fmt.Sprintf("hypershift.openshift.io/infra-id=%s", clusterName),
+	})
 	if err != nil {
-		return err
-	}
-	mc := &clusterv1.ManagedCluster{}
-	if err := runtime.DefaultUnstructuredConverter.FromUnstructured(mcu.UnstructuredContent(), mc); err != nil {
-		return err
+		return nil, err
 	}
 
-	var consoleURL string
-	for _, c := range mc.Status.ClusterClaims {
-		if c.Name == ConsoleURLClusterClaim {
-			consoleURL = c.Value
-		}
+	hd := &hypershiftdeploymentv1alpha1.HypershiftDeployment{}
+	if err := runtime.DefaultUnstructuredConverter.FromUnstructured(hdus.Items[0].UnstructuredContent(), hd); err != nil {
+		return nil, err
 	}
-	if consoleURL == "" {
-		return fmt.Errorf("console url not found for cluster %s", clusterName)
-	}
-
-	return helpers.Openbrowser(consoleURL)
+	return hd, nil
 }